Integrate from book source
Remove Ray::depth; pass it through in the places where it is needed.
Many variable name and small code cleanups.
BSSRDF improvements
More C++11isms
Clean up RNG interfaces
Fix subtle bug in LatinHypercube() sample shuffling

+++ src/core/pbrtparse.cpp
@@ -154,7 +154,7 @@ static void ArrayFree(ParamArray *ra) {
 
 
 static void FreeArgs() {
-    for (uint32_t i = 0; i < cur_paramlist.size(); ++i)
+    for (size_t i = 0; i < cur_paramlist.size(); ++i)
         free((char *)cur_paramlist[i].arg);
     cur_paramlist.erase(cur_paramlist.begin(), cur_paramlist.end());
 }
@@ -1776,9 +1776,9 @@ yyreduce:
 /* Line 1807 of yacc.c  */
 #line 313 "/Users/mpharr/pbrt-3ed/src/core/pbrtparse.yy"
     {
-    for (uint32_t i = 0; i < cur_paramlist.size(); ++i) {
+    for (size_t i = 0; i < cur_paramlist.size(); ++i) {
         if (cur_paramlist[i].isString) {
-            for (uint32_t j = 0; j < (uint32_t)cur_paramlist[i].size; ++j)
+            for (size_t j = 0; j < cur_paramlist[i].size; ++j)
                 free(((char **)cur_paramlist[i].arg)[j]);
         }
     }
@@ -2445,7 +2445,7 @@ static const char *paramTypeToName(int type) {
 
 static void InitParamSet(ParamSet &ps, SpectrumType type) {
     ps.Clear();
-    for (uint32_t i = 0; i < cur_paramlist.size(); ++i) {
+    for (size_t i = 0; i < cur_paramlist.size(); ++i) {
         int type;
         std::string name;
         if (lookupType(cur_paramlist[i].name, &type, name)) {
